requirements.txt
asgiref==3.9.2
Brotli==1.1.0
cffi==2.0.0
charset-normalizer==3.4.3
cssselect2==0.8.0
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.60.1
pillow==11.3.0
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-decouple==3.8
python-dotenv==1.1.1
reportlab==4.4.4
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
